pyproject.toml
[project]
name = "ai-stock-scheduler"
version = "1.0.0"
description = "AI Stock Scheduler - 定时任务调度服务（独立部署）"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.27.0",
    "gunicorn>=21.2.0",
    "openai>=1.0.0",
    "structlog>=25.3.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "pymysql>=1.1.0",
    "aiomysql>=0.2.0",
    "redis[hiredis]>=5.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "akshare>=1.14.0",
    "pandas>=2.0.0",
    "numpy>=1.26,<2.0",
    "requests>=2.31.0",
]
